chore: refactor component placements
