+++ src/components/probe/index.ts
@@ -22,10 +22,10 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
-import { ValidateResponseStatus } from './../../utils/alert'
-import { processProbeStatus } from '../../utils/process-server-status'
-import { probing } from '../../utils/probing'
-import { validateResponse } from '../../utils/alert'
+import { ValidateResponseStatus } from '../notification/alert'
+import { processProbeStatus } from '../notification/process-server-status'
+import { probing } from './probing'
+import { validateResponse } from '../notification/alert'
 import { Probe } from '../../interfaces/probe'
 import { Notification } from '../../interfaces/notification'
 import { probeLog } from '../logger'
